增加获取球员接口

+++ frame-dao/src/main/java/com/frame/dao/UserLoginDao.java
@@ -1,9 +1,10 @@
 package com.frame.dao;
 
+import java.util.List;
+
 import com.frame.dao.base.BaseDao;
+import com.frame.domain.User;
 import com.frame.domain.UserLogin;
 
-public interface UserLoginDao extends BaseDao<UserLogin, Long> {
-	
-
+public interface UserLoginDao extends BaseDao<UserLogin, Integer> {
 }
