+++ frame-service/src/main/java/com/frame/service/UserLoginService.java
@@ -1,9 +1,11 @@
 package com.frame.service;
 
+import java.util.List;
+
+import com.frame.domain.User;
 import com.frame.domain.UserLogin;
 import com.frame.service.base.BaseService;
 
-public interface UserLoginService extends BaseService<UserLogin, Long> {
-	
+public interface UserLoginService extends BaseService<UserLogin, Integer> {
 	
 }
